Code improvements regarding VBW/RBW calculation.

+++ src/morrow/sa.h
@@ -21,7 +21,7 @@ int32_t SetInterfaceType(SET9052 *a1, /*int16_t*/InterfaceEnum type, int32_t a3)
 int32_t SetCellMode(SET9052 *a1, int16_t mode);
 int32_t SetPortNum(SET9052 *a1, uint16_t a2) ;
 int32_t GetRBWwide(int16_t value);
-int32_t VBWFreqFromCode(int16_t a1);
+float64_t /*int32_t*/ VBWFreqFromCode(int16_t a1);
 float64_t /*int32_t*/ RBWFreqFromCode(int16_t a1);
 int32_t DefltSetTimeRBW(int16_t timeValue);
 int32_t DefltSetTimeVBW(int16_t timeValue);
